Add markdown to shorten description
