+++ feature-governance-impl/src/main/java/io/novafoundation/nova/feature_governance_impl/presentation/referenda/details/di/ReferendumDetailsModule.kt
@@ -1,5 +1,7 @@
 package io.novafoundation.nova.feature_governance_impl.presentation.referenda.details.di
 
+import android.content.Context
+import android.text.util.Linkify
 import androidx.fragment.app.Fragment
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
@@ -7,6 +9,8 @@ import dagger.Module
 import dagger.Provides
 import dagger.multibindings.IntoMap
 import io.noties.markwon.Markwon
+import io.noties.markwon.ext.strikethrough.StrikethroughPlugin
+import io.noties.markwon.linkify.LinkifyPlugin
 import io.novafoundation.nova.common.address.AddressIconGenerator
 import io.novafoundation.nova.common.di.viewmodel.ViewModelKey
 import io.novafoundation.nova.common.di.viewmodel.ViewModelModule
@@ -17,6 +21,7 @@ import io.novafoundation.nova.feature_governance_api.domain.referendum.details.R
 import io.novafoundation.nova.feature_governance_impl.data.GovernanceSharedState
 import io.novafoundation.nova.feature_governance_impl.domain.dapp.GovernanceDAppsInteractor
 import io.novafoundation.nova.feature_governance_impl.domain.identity.GovernanceIdentityProviderFactory
+import io.novafoundation.nova.feature_governance_impl.markdown.StylePlugin
 import io.novafoundation.nova.feature_governance_impl.presentation.GovernanceRouter
 import io.novafoundation.nova.feature_governance_impl.presentation.referenda.common.ReferendumFormatter
 import io.novafoundation.nova.feature_governance_impl.presentation.referenda.details.ReferendumDetailsPayload
@@ -25,6 +30,15 @@ import io.novafoundation.nova.feature_wallet_api.domain.TokenUseCase
 
 @Module(includes = [ViewModelModule::class])
 class ReferendumDetailsModule {
+
+    @Provides
+    fun provideMarkwon(context: Context): Markwon {
+        return Markwon.builder(context)
+            .usePlugin(LinkifyPlugin.create(Linkify.EMAIL_ADDRESSES or Linkify.WEB_URLS))
+            .usePlugin(StylePlugin(context))
+            .usePlugin(StrikethroughPlugin.create())
+            .build()
+    }
 
     @Provides
     @IntoMap
