+++ feature-governance-impl/src/main/java/io/novafoundation/nova/feature_governance_impl/presentation/referenda/details/model/ShortenedTextModel.kt
@@ -1,17 +1,19 @@
 package io.novafoundation.nova.feature_governance_impl.presentation.referenda.details.model
 
+import android.text.TextUtils
+
 private const val ELLIPSIS = "..."
 
-class ShortenedTextModel private constructor(val shortenedText: String, val hasMore: Boolean) {
+class ShortenedTextModel private constructor(val shortenedText: CharSequence, val hasMore: Boolean) {
 
     companion object {
-        fun from(text: String, characterLimit: Int): ShortenedTextModel {
+        fun from(text: CharSequence, characterLimit: Int): ShortenedTextModel {
             require(characterLimit >= ELLIPSIS.length)
 
             return if (text.length > characterLimit) {
-                val shortened = text.substring(0, characterLimit - ELLIPSIS.length) + ELLIPSIS
-
-                ShortenedTextModel(shortenedText = shortened, hasMore = true)
+                val shortened = text.subSequence(0, characterLimit - ELLIPSIS.length)
+                val shortenedEllipsized = TextUtils.concat(shortened, ELLIPSIS)
+                ShortenedTextModel(shortenedText = shortenedEllipsized, hasMore = true)
             } else {
                 ShortenedTextModel(shortenedText = text, hasMore = false)
             }
